Revert "autotools: try improving thread detection."
This reverts commit 14c6378fc41906db14464e17aadc7d94a825fafd.

So either we get to fix T880 or T986. I do think the problem is more tricky than
that, so I am for the moment going to get T880 broken rather than T986 as I think
the later will show up in more case for our user base. Will try to find a proper
solution for everyone before the next release.