Emotion VLC: remove quite all the old hacks.
This commit remove all the hacks that was there on the loading stage.
Instead of using the tmp format callback system now simply connect to
the MediaPlayerPlaying event on first startup and pause the video as
soon as it start.

No more need for the tmp buffer allocation and the delayed stop using idler.

All this fix the first outstanding bug: audio was audible on start (before the real
play command from emotion)