Evas filters: Implement mapped blend for alpha buffers
Allow repeat fillmode in blend() for:
alpha --> alpha
alpha --> rgba
rgba --> alpha

Alpha scaling is not implemented yet, but it is not actually
required. Indeed, only proxies can have a different size and
proxies are RGBA images, not alpha.

Alpha scaling may or may not become a requirement in the future,
or for other purposes, but not yet.