Emotion VLC: fix the random-player-stop bug !!
Bug explanation:
VLC has 2 different async mechanism that we use, one is the event callback
system that report player status changes and one is the display func that
is called at every frame. They run independently and can occur "at the same time".
We use ecore_pipe_write() in both and the result is that occasionally we send to
the pipe 2 commands mixed togheter (the FRAME_NEW and the LENGTH_CHANGED).
So emotion (on the other side of the pipe) receive a wrong command and stop the player.

The fix:
Use a mutex (Eina_Lock) while sending commands to the pipe so that we can send 2
commands "at the same time".

The SPANK credits:
BIG BIG SPANK goes to cedric (of course) that removed the command lock mechanism in
revision 10ebda80419