Evas: Fix build issue when building without drm
@bugfix: Fix issue of building efl without drm or drm hw accel options

By default drm & drm hw acceleration is off as an option. This caused
a problem when passing no options to autogen due to a missing define.
This commit fixes that mess by only defining HAVE_DRM_HW_ACCEL If we
actually have the deps for it, rather than using an AM_CONDITIONAL as
it did previously.

Signed-off-by: Chris Michael <cp.michael@samsung.com>