autotools: let's be precise on how we do support ABI stability.
NOTE: We do use enum in structure. There size can be changed by changing
the parameter of the compiler, this obviously will break ABI. As long as
you use the same compiler (and normally any future version of that compiler)
with the same set of CFLAGS/CXXFLAGS, you will have ABI stability.