[Elm_Entry] Update selection handlers position and its visibility when scrolling
Summary:
Selection handlers' position are not updated when entry is scrolled.
You can check on elementary_test -> Entry 3 (or Entry 4 which is scrollable) -> Select some text -> Scroll entry and see.
This patch is submitted to fix this bug. It update selection handlers' position when entry is scrolled and hide them if needed.
@fix

Reviewers: raster

Differential Revision: https://phab.enlightenment.org/D640