edje_edit: function that will return the source code of the loaded edje edit object.
Summary:
This function will return the source code of the edje edit object that is given to that function.
Also, that function will collect all resources required by this group.
That means all images, fonts, styles, data and color_classes that is mentioned
and set in descriptions of parts in this group (that is represented by given
Edje Edit object).
@feature

Added function for generating source code:
- edje_edit_source_generate(Evas_Object *obj);

Reviewers: cedric, seoz, raster

Reviewed By: raster

CC: reutskiy.v.v, cedric

Differential Revision: https://phab.enlightenment.org/D649