Workaround broken popup on Terminology
As the comment in the code explains, when ctxpopup is shown, it's giving
focus to the inner list, which immediately decides to select the first
of its items, calling, in the process, its selection callback. This has
the very unfortunate, annoying and uselessly broken effect of making the
popup never showing up, and instead triggering the first action in it.

Since this behavior of "select first item on focus" seems to be the path
taken in other places, I'm letting the people that know what they want
to do about that figure out how to fix it properly.