autotools: fix to many argument on shell script during make check
The autogenerated check macro by autotools lead to a huge number of file
being added on the command line of a script, when the only things it does
is to test for the definition of those strings.

Hoping this is going to fix the problem. If that doesn't we will need to write
our own CHECK macro here.