add hand1 cursor theme element and ensure elm cursor code actually works
this adds a hand1 sample cursor theme/image so the theme can find
cursors, (more should be added later), and now actually makes the elm
cursor handling properly set the cursor and update the hotx/y as
needed by tracking the hot swallow object. this relies on a fix in efl
as well.