rg_etc: Fix debug build
In 4053911e we tried to fix the debug build and failed. The function API
actually asks for coords as first parameter. Our nightly builds expose
this debug build and have been failing due to this.

It also makes clear that the debug part of this code was never really
used upstream...