build: Fix coevrage build with eolian and lcov 1.10
Mixing up the new eolian test with its generated eo_lexer code and a new
version of lcov like 1.10. It works fine with lcov 1.6.

The problem is that this combination somhow triggers a wrong file path
with filename NONE. It was only triggered when adding the eolian test
suite so its a combination here.

For now we use this bandaid to remove the broken files. Need to find a
way to report this to lcov upstream.