Eo: Make it clearer that a NULL deref can't happen.
We check _current_table for NULL, and then populate it (it's a global)
through another function, but we don't really check it's not NULL before
using it, we just assume because of an indirect other variable.

This confused coveritiy, can confuse humans too, and in general risky
(if something changes).

CID 1039419