Evas TGV: Add ETC2 support to the TGV saver
This will call the new ETC2 encoding functions.
Note that the quality and performance will be horrible, but at
least alpha should be supported.

Also, there is no way to choose between ETC1 and ETC2 from
the client API side, which, well, sucks. So ETC2 is selected
if and only if the image has alpha (according to its flag).

@feature: Encode images in ETC2 with support for Alpha (if needed).