Revert "Eolian CXX: Fix make distcheck"
This is essentially a revert of my recent commit. This is wrong. These
files are not built sources, as they are not/should not be used by
anything else. They are just generated files. Putting them there
essentially makes the build linear and just wrong.

See commit 2010de6810ea31116c77f2e08eb0dc9dc7034768 which first
introduced this change for more details.

This reverts commit 43038144c27af0222b386f67b565401a283877e5.