edje: Edje_Edit - add edje_edit_sound_sample_add() function that allows user to add new sound sample to collection
Summary:
New function provides the ability to add new sound samples to currently loaded collection.
It takes path to the sound-file and name that will define this sound in collection as parameters.
The sound-file of created sample will be imported to .edj-file. Due to the fact that the
edje_import_font_file() function performs the same actions as we need to perform for importing
sound file, it is renamed to _edje_edit_file_import() that takes one additional
parameter that specifies the compression level for eet_write(). This renaming is done to
avoid code duplication.
Also this commit provides small fix that shifts all the items of arrays of tones and samples
after deletion of tone or sample correspondingly. This fix simplifies the calculation of
id of item on adding new sample or tone. Also it makes generated .edc-code look better.

@feature

Reviewers: cedric, Hermet, seoz, raster

CC: reutskiy.v.v, cedric

Differential Revision: https://phab.enlightenment.org/D1014

Signed-off-by: Cedric BAIL <c.bail@partner.samsung.com>