Revert "Coverage: Fix base directory."
This reverts commit 842e8e9fa01f2ce0c217dec2578d5ee3e0911a8b.

Scratching my head over this for a long time now. It sneaked in when
jenkins nightly builds which generates coverage reports was broken
due to newer gettext error handling. After that was fixed the build
kept being broken but now in coverage generation.

Finnaly found this change. Why was it done? Did you actually test it?
Coverage generation worked fine on my local system as well as on
jenkins when I worked on this. If it is broken for you we might need
to have another look, but not by breaking jenkins.